Revert "Core/Spells: fixed packet structure for spell_start and spell_go packets when ammo was involved"
This reverts commit c9baaec657a248179d10945011051f24d4378e8d.

+++ src/server/game/Server/Packets/SpellPackets.cpp
@@ -45,7 +45,7 @@ ByteBuffer& operator<<(ByteBuffer& data, WorldPackets::Spells::MissileTrajectory
 ByteBuffer& operator<<(ByteBuffer& data, WorldPackets::Spells::SpellAmmo const& spellAmmo)
 {
     data << int32(spellAmmo.DisplayID);
-    data << uint8(spellAmmo.InventoryType);
+    data << int32(spellAmmo.InventoryType);
     return data;
 }
